feat: combining preference survey and log out into beatiful profile page
