+++ mobile-app/lib/src/home/home_page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:golocal/src/event/ui/events_view_page.dart';
-import 'package:golocal/src/user/profile_page.dart';
-import 'package:golocal/src/preference_survey/ui/preference_survey_page.dart';
+import 'package:golocal/src/user/ui/profile_page.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -15,7 +14,6 @@ class _HomePageState extends State<HomePage> {
     const Text('Map'),
     EventsViewPage(),
     ProfilePage(),
-    PreferenceSurveyPage(),
   ];
 
   @override
@@ -32,16 +30,10 @@ class _HomePageState extends State<HomePage> {
               const Tab(icon: Icon(Icons.map), text: 'Map'),
               const Tab(icon: Icon(Icons.event), text: 'Events'),
               const Tab(icon: Icon(Icons.person), text: 'Profile'),
-              const Tab(icon: Icon(Icons.settings), text: 'Settings'),
             ],
           ),
           body: TabBarView(
-            children: [
-              Center(child: Text('Map')),
-              EventsViewPage(),
-              ProfilePage(),
-              PreferenceSurveyPage(),
-            ],
+            children: pages,
           ),
         ),
       );
