+++ mobile-app/lib/src/user/ui/profile_page.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:golocal/src/auth/bloc/auth_bloc.dart';
+import 'package:golocal/src/preference_survey/ui/preference_survey_page.dart';
+
+class ProfilePage extends StatelessWidget {
+  const ProfilePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          'Profile Page',
+          style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24),
+        ),
+        centerTitle: true,
+      ),
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          ElevatedButton(
+            onPressed: () {
+              context.read<AuthBloc>().add(const AuthLogout());
+            },
+            child: const Text('Log out'),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (context) => PreferenceSurveyPage()),
+              );
+            },
+            child: const Text('Fill out the preference survey'),
+          ),
+        ],
+      ),
+    );
+  }
+}
